+++ packages/registry/src/index.ts
@@ -2,48 +2,55 @@ import {
   type ActionParams,
   ActionType,
   type BridgeActionParams,
-  PluginActionNotImplementedError,
+  type DelegateActionParams,
   type MintActionParams,
   type OptionsActionParams,
-  type SwapActionParams,
+  PluginActionNotImplementedError,
   type QuestActionParams,
-  type DelegateActionParams,
-  type TransactionFilter,
   type StakeActionParams,
+  type SwapActionParams,
+  type TransactionFilter,
+  type VoteActionParams,
 } from '@rabbitholegg/questdk'
 
-import { Connext } from '@rabbitholegg/questdk-plugin-connext'
-import { Uniswap } from '@rabbitholegg/questdk-plugin-uniswap'
-import { Stargate } from '@rabbitholegg/questdk-plugin-stargate'
 import { Across } from '@rabbitholegg/questdk-plugin-across'
-import { Polygon } from '@rabbitholegg/questdk-plugin-polygon'
-import { Optimism } from '@rabbitholegg/questdk-plugin-optimism'
-import { Hop } from '@rabbitholegg/questdk-plugin-hop'
 import { Arbitrum } from '@rabbitholegg/questdk-plugin-arbitrum'
-import { GMX } from '@rabbitholegg/questdk-plugin-gmx'
-import { Camelot } from '@rabbitholegg/questdk-plugin-camelot'
-import { Tally } from '@rabbitholegg/questdk-plugin-tally'
-import { BasePaint } from '@rabbitholegg/questdk-plugin-basepaint'
-import { Hyphen } from '@rabbitholegg/questdk-plugin-hyphen'
-import { Paraswap } from '@rabbitholegg/questdk-plugin-paraswap'
-import { Rabbithole } from '@rabbitholegg/questdk-plugin-rabbithole'
-import { Symbiosis } from '@rabbitholegg/questdk-plugin-symbiosis'
-import { OkuTrade } from '@rabbitholegg/questdk-plugin-okutrade'
-import { Zora } from '@rabbitholegg/questdk-plugin-zora'
 import { Balancer } from '@rabbitholegg/questdk-plugin-balancer'
-import { TraderJoe } from '@rabbitholegg/questdk-plugin-traderjoe'
-import { Synapse } from '@rabbitholegg/questdk-plugin-synapse'
-import { WooFi } from '@rabbitholegg/questdk-plugin-woofi'
-import { Sushi } from '@rabbitholegg/questdk-plugin-sushi'
-import { Treasure } from '@rabbitholegg/questdk-plugin-treasure'
-import { Pendle } from '@rabbitholegg/questdk-plugin-pendle'
+import { BasePaint } from '@rabbitholegg/questdk-plugin-basepaint'
+import { Boost } from '@rabbitholegg/questdk-plugin-boost'
+import { Camelot } from '@rabbitholegg/questdk-plugin-camelot'
+import { Connext } from '@rabbitholegg/questdk-plugin-connext'
+import { GMX } from '@rabbitholegg/questdk-plugin-gmx'
 import { HandleFi } from '@rabbitholegg/questdk-plugin-handlefi'
+import { Hop } from '@rabbitholegg/questdk-plugin-hop'
+import { Hyphen } from '@rabbitholegg/questdk-plugin-hyphen'
+import { Llama } from '@rabbitholegg/questdk-plugin-llama'
 import { Mirror } from '@rabbitholegg/questdk-plugin-mirror'
-import { Soundxyz } from '@rabbitholegg/questdk-plugin-soundxyz'
-import { Vela } from '@rabbitholegg/questdk-plugin-vela'
 import { Mux } from '@rabbitholegg/questdk-plugin-mux'
+import { OkuTrade } from '@rabbitholegg/questdk-plugin-okutrade'
+import { Optimism } from '@rabbitholegg/questdk-plugin-optimism'
+import { Paraswap } from '@rabbitholegg/questdk-plugin-paraswap'
+import { Pendle } from '@rabbitholegg/questdk-plugin-pendle'
+import { Polygon } from '@rabbitholegg/questdk-plugin-polygon'
+import { Rabbithole } from '@rabbitholegg/questdk-plugin-rabbithole'
+import { Soundxyz } from '@rabbitholegg/questdk-plugin-soundxyz'
+import { Stargate } from '@rabbitholegg/questdk-plugin-stargate'
+import { Sushi } from '@rabbitholegg/questdk-plugin-sushi'
+import { Symbiosis } from '@rabbitholegg/questdk-plugin-symbiosis'
+import { Synapse } from '@rabbitholegg/questdk-plugin-synapse'
+import { Tally } from '@rabbitholegg/questdk-plugin-tally'
+import { TraderJoe } from '@rabbitholegg/questdk-plugin-traderjoe'
+import { Treasure } from '@rabbitholegg/questdk-plugin-treasure'
+import { Uniswap } from '@rabbitholegg/questdk-plugin-uniswap'
+import { Vela } from '@rabbitholegg/questdk-plugin-vela'
+import { WooFi } from '@rabbitholegg/questdk-plugin-woofi'
+import { Zora } from '@rabbitholegg/questdk-plugin-zora'
 import { ENTRYPOINT } from './contract-addresses'
-import type { IntentParams, MintIntentParams, IActionPlugin } from '../../utils/src/types/'
+import type {
+  IntentParams,
+  MintIntentParams,
+  IActionPlugin,
+} from '../../utils/src/types/'
 
 export const plugins: Record<string, IActionPlugin> = {
   [Connext.pluginId]: Connext,
@@ -76,6 +83,8 @@ export const plugins: Record<string, IActionPlugin> = {
   [Soundxyz.pluginId]: Soundxyz,
   [Mux.pluginId]: Mux,
   [Vela.pluginId]: Vela,
+  [Boost.pluginId]: Boost,
+  [Llama.pluginId]: Llama,
 }
 
 export const getPlugin = (pluginId: string) => {
@@ -89,7 +98,7 @@ export const getPlugin = (pluginId: string) => {
 export const getTxIntent = (
   plugin: IActionPlugin,
   actionType: ActionType,
-  params: IntentParams
+  params: IntentParams,
 ) => {
   switch (actionType) {
     case ActionType.Mint:
@@ -110,11 +119,17 @@ export const executePlugin = (
 ): Promise<TransactionFilter | PluginActionNotImplementedError> => {
   switch (actionType) {
     case ActionType.Bridge:
-      return plugin.bridge(params as unknown as BridgeActionParams)
+      if (plugin.bridge === undefined) {
+        return Promise.reject(new PluginActionNotImplementedError())
+      } else return plugin.bridge(params as unknown as BridgeActionParams)
     case ActionType.Swap:
-      return plugin.swap(params as unknown as SwapActionParams)
+      if (plugin.swap === undefined) {
+        return Promise.reject(new PluginActionNotImplementedError())
+      } else return plugin.swap(params as unknown as SwapActionParams)
     case ActionType.Mint:
-      return plugin.mint(params as unknown as MintActionParams)
+      if (plugin.mint === undefined) {
+        return Promise.reject(new PluginActionNotImplementedError())
+      } else return plugin.mint(params as unknown as MintActionParams)
     case ActionType.Delegate: {
       if (plugin.delegate === undefined) {
         return Promise.reject(new PluginActionNotImplementedError())
@@ -134,6 +149,11 @@ export const executePlugin = (
       if (plugin.options === undefined) {
         return Promise.reject(new PluginActionNotImplementedError())
       } else return plugin.options(params as unknown as OptionsActionParams)
+    }
+    case ActionType.Vote: {
+      if (plugin.vote === undefined) {
+        return Promise.reject(new PluginActionNotImplementedError())
+      } else return plugin.vote(params as unknown as VoteActionParams)
     }
     default:
       throw new Error(`Unknown action type "${actionType}"`)