+++ packages/treasure/src/index.ts
@@ -5,6 +5,7 @@ import {
 
 import {
   mint,
+  stake,
   swap,
   getSupportedChainIds,
   getSupportedTokenAddresses,
@@ -15,6 +16,7 @@ export const Treasure: IActionPlugin = {
   getSupportedTokenAddresses,
   getSupportedChainIds,
   mint,
+  stake,
   swap,
   bridge: async () => new PluginActionNotImplementedError(),
 }