+++ packages/utils/src/constants/index.ts
@@ -1,1 +1,2 @@
-export * from './contract-addresses'
+export { CHAIN_TO_TOKENS } from './contract-addresses'
+export { Chains } from './chain-ids'