+++ packages/vela/src/chain-ids.ts
@@ -1,3 +1,3 @@
-import { Chains } from './utils'
+import { Chains } from '@rabbitholegg/questdk-plugin-utils'
 
 export const CHAIN_ID_ARRAY = [Chains.ARBITRUM_ONE]