+++ packages/balancer/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @rabbitholegg/questdk-plugin-balancer
 
+## 1.0.0-alpha.13
+
+### Minor Changes
+
+- [#206](https://github.com/rabbitholegg/questdk-plugins/pull/206) [`5676053`](https://github.com/rabbitholegg/questdk-plugins/commit/5676053599e47b8888a243c2b5385b322ae206d1) Thanks [@mmackz](https://github.com/mmackz)! - use utils package
+
+## 1.0.0-alpha.12
+
+### Patch Changes
+
+- Updated dependencies [[`30c9913`](https://github.com/rabbitholegg/questdk-plugins/commit/30c99133337b9986d28081baad40186285ae0559)]:
+  - @rabbitholegg/questdk-plugin-utils@1.0.0-alpha.5
+
+## 1.0.0-alpha.11
+
+### Patch Changes
+
+- Updated dependencies [[`9f081c7`](https://github.com/rabbitholegg/questdk-plugins/commit/9f081c7e89479aa8e29abd8d36908515dfc16e3b), [`85c8cd8`](https://github.com/rabbitholegg/questdk-plugins/commit/85c8cd82070ba9a17f5181183533b4f1a5b117d7), [`9fe746a`](https://github.com/rabbitholegg/questdk-plugins/commit/9fe746ad29ba042d40dc8ab610cd531c83a878c3)]:
+  - @rabbitholegg/questdk-plugin-utils@1.0.0-alpha.4
+
 ## 1.0.0-alpha.10
 
 ### Patch Changes