+++ packages/mux/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @rabbitholegg/questdk-plugin-mux
 
+## 1.0.0-alpha.2
+
+### Minor Changes
+
+- [#199](https://github.com/rabbitholegg/questdk-plugins/pull/199) [`e934b2d`](https://github.com/rabbitholegg/questdk-plugins/commit/e934b2da5427e038268d803f482df9c52dd052fe) Thanks [@Quazia](https://github.com/Quazia)! - Move questdk dependencies to root
+
 ## 1.0.0-alpha.1
 
 ### Minor Changes