+++ packages/registry/CHANGELOG.md
@@ -1,5 +1,99 @@
 # @rabbitholegg/questdk-plugin-registry
 
+## 1.0.0-alpha.62
+
+### Patch Changes
+
+- Updated dependencies [[`14ab4a0`](https://github.com/rabbitholegg/questdk-plugins/commit/14ab4a0f536ab7a21f3502faf57a7ee2404b08e1)]:
+  - @rabbitholegg/questdk-plugin-okutrade@1.0.0-alpha.17
+
+## 1.0.0-alpha.61
+
+### Patch Changes
+
+- Updated dependencies [[`88d3496`](https://github.com/rabbitholegg/questdk-plugins/commit/88d349651a4aba3d75fd27371eb97878974df860), [`2341dd7`](https://github.com/rabbitholegg/questdk-plugins/commit/2341dd7c8cebc454ac0afc413c7fa123714ecdbd)]:
+  - @rabbitholegg/questdk-plugin-vela@1.0.0-alpha.4
+  - @rabbitholegg/questdk-plugin-gmx@1.0.0-alpha.22
+
+## 1.0.0-alpha.60
+
+### Patch Changes
+
+- Updated dependencies [[`5676053`](https://github.com/rabbitholegg/questdk-plugins/commit/5676053599e47b8888a243c2b5385b322ae206d1), [`415e400`](https://github.com/rabbitholegg/questdk-plugins/commit/415e400e0d612cfab5d52d7c9a31489c6e8694d2), [`22dc15b`](https://github.com/rabbitholegg/questdk-plugins/commit/22dc15b4fa8618da325011f433919a303fa9ca9d)]:
+  - @rabbitholegg/questdk-plugin-balancer@1.0.0-alpha.13
+  - @rabbitholegg/questdk-plugin-camelot@1.0.0-alpha.17
+  - @rabbitholegg/questdk-plugin-uniswap@1.0.0-alpha.17
+
+## 1.0.0-alpha.59
+
+### Patch Changes
+
+- [#183](https://github.com/rabbitholegg/questdk-plugins/pull/183) [`af36abe`](https://github.com/rabbitholegg/questdk-plugins/commit/af36abecef9a95861fab8f918aaac330640c24fd) Thanks [@gnomadic](https://github.com/gnomadic)! - Introducing Llama Plugin with Vote Action
+
+- Updated dependencies [[`af36abe`](https://github.com/rabbitholegg/questdk-plugins/commit/af36abecef9a95861fab8f918aaac330640c24fd), [`6a8bf0d`](https://github.com/rabbitholegg/questdk-plugins/commit/6a8bf0d629161c9da279c020370da4cdeeaf68d2), [`2d042c7`](https://github.com/rabbitholegg/questdk-plugins/commit/2d042c7f0f45665a631f16c3e7870eb76968c138), [`6fa08fd`](https://github.com/rabbitholegg/questdk-plugins/commit/6fa08fdb856dd08454a304f30e6c886dec36f6dc)]:
+  - @rabbitholegg/questdk-plugin-llama@1.0.0-alpha.9
+  - @rabbitholegg/questdk-plugin-handlefi@1.0.0-alpha.5
+  - @rabbitholegg/questdk-plugin-uniswap@1.0.0-alpha.16
+  - @rabbitholegg/questdk-plugin-vela@1.0.0-alpha.3
+  - @rabbitholegg/questdk-plugin-across@1.0.0-alpha.14
+  - @rabbitholegg/questdk-plugin-balancer@1.0.0-alpha.12
+  - @rabbitholegg/questdk-plugin-basepaint@1.0.0-alpha.16
+  - @rabbitholegg/questdk-plugin-boost@1.0.0-alpha.2
+  - @rabbitholegg/questdk-plugin-camelot@1.0.0-alpha.16
+  - @rabbitholegg/questdk-plugin-connext@1.0.0-alpha.15
+  - @rabbitholegg/questdk-plugin-gmx@1.0.0-alpha.21
+  - @rabbitholegg/questdk-plugin-hop@1.0.0-alpha.14
+  - @rabbitholegg/questdk-plugin-hyphen@1.0.0-alpha.14
+  - @rabbitholegg/questdk-plugin-okutrade@1.0.0-alpha.16
+  - @rabbitholegg/questdk-plugin-optimism@1.0.0-alpha.14
+  - @rabbitholegg/questdk-plugin-paraswap@1.0.0-alpha.20
+  - @rabbitholegg/questdk-plugin-polygon@1.0.0-alpha.14
+  - @rabbitholegg/questdk-plugin-rabbithole@1.0.0-alpha.8
+  - @rabbitholegg/questdk-plugin-stargate@1.0.0-alpha.20
+  - @rabbitholegg/questdk-plugin-symbiosis@1.0.0-alpha.15
+  - @rabbitholegg/questdk-plugin-synapse@1.0.0-alpha.8
+  - @rabbitholegg/questdk-plugin-tally@1.0.0-alpha.12
+  - @rabbitholegg/questdk-plugin-traderjoe@1.0.0-alpha.14
+
+## 1.0.0-alpha.58
+
+### Minor Changes
+
+- [#198](https://github.com/rabbitholegg/questdk-plugins/pull/198) [`d29da86`](https://github.com/rabbitholegg/questdk-plugins/commit/d29da863bdf5ea55d3a15fdc80c37826967d2d64) Thanks [@mmackz](https://github.com/mmackz)! - add boostpass mint action to boost plugin
+
+### Patch Changes
+
+- Updated dependencies [[`c144168`](https://github.com/rabbitholegg/questdk-plugins/commit/c14416849b02ca373fbf75e72b8e23301a98aa56), [`e934b2d`](https://github.com/rabbitholegg/questdk-plugins/commit/e934b2da5427e038268d803f482df9c52dd052fe), [`d29da86`](https://github.com/rabbitholegg/questdk-plugins/commit/d29da863bdf5ea55d3a15fdc80c37826967d2d64)]:
+  - @rabbitholegg/questdk-plugin-treasure@1.0.0-alpha.5
+  - @rabbitholegg/questdk-plugin-handlefi@1.0.0-alpha.4
+  - @rabbitholegg/questdk-plugin-soundxyz@1.0.0-alpha.4
+  - @rabbitholegg/questdk-plugin-mirror@1.0.0-alpha.4
+  - @rabbitholegg/questdk-plugin-pendle@1.0.0-alpha.9
+  - @rabbitholegg/questdk-plugin-sushi@1.0.0-alpha.4
+  - @rabbitholegg/questdk-plugin-woofi@1.0.0-alpha.11
+  - @rabbitholegg/questdk-plugin-vela@1.0.0-alpha.2
+  - @rabbitholegg/questdk-plugin-mux@1.0.0-alpha.2
+  - @rabbitholegg/questdk-plugin-boost@1.0.0-alpha.1
+  - @rabbitholegg/questdk-plugin-across@1.0.0-alpha.13
+  - @rabbitholegg/questdk-plugin-balancer@1.0.0-alpha.11
+  - @rabbitholegg/questdk-plugin-basepaint@1.0.0-alpha.15
+  - @rabbitholegg/questdk-plugin-camelot@1.0.0-alpha.15
+  - @rabbitholegg/questdk-plugin-connext@1.0.0-alpha.14
+  - @rabbitholegg/questdk-plugin-gmx@1.0.0-alpha.20
+  - @rabbitholegg/questdk-plugin-hop@1.0.0-alpha.13
+  - @rabbitholegg/questdk-plugin-hyphen@1.0.0-alpha.13
+  - @rabbitholegg/questdk-plugin-okutrade@1.0.0-alpha.15
+  - @rabbitholegg/questdk-plugin-optimism@1.0.0-alpha.13
+  - @rabbitholegg/questdk-plugin-paraswap@1.0.0-alpha.19
+  - @rabbitholegg/questdk-plugin-polygon@1.0.0-alpha.13
+  - @rabbitholegg/questdk-plugin-rabbithole@1.0.0-alpha.7
+  - @rabbitholegg/questdk-plugin-stargate@1.0.0-alpha.19
+  - @rabbitholegg/questdk-plugin-symbiosis@1.0.0-alpha.14
+  - @rabbitholegg/questdk-plugin-synapse@1.0.0-alpha.7
+  - @rabbitholegg/questdk-plugin-tally@1.0.0-alpha.11
+  - @rabbitholegg/questdk-plugin-traderjoe@1.0.0-alpha.13
+  - @rabbitholegg/questdk-plugin-uniswap@1.0.0-alpha.15
+
 ## 1.0.0-alpha.57
 
 ### Patch Changes