+++ packages/vela/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @rabbitholegg/questdk-plugin-vela
 
+## 1.0.0-alpha.4
+
+### Minor Changes
+
+- [#192](https://github.com/rabbitholegg/questdk-plugins/pull/192) [`88d3496`](https://github.com/rabbitholegg/questdk-plugins/commit/88d349651a4aba3d75fd27371eb97878974df860) Thanks [@mmackz](https://github.com/mmackz)! - add support for advanced TPSL orders
+
+## 1.0.0-alpha.3
+
+### Minor Changes
+
+- [#174](https://github.com/rabbitholegg/questdk-plugins/pull/174) [`6fa08fd`](https://github.com/rabbitholegg/questdk-plugins/commit/6fa08fdb856dd08454a304f30e6c886dec36f6dc) Thanks [@lovetrading10](https://github.com/lovetrading10)! - add VLP staking to vela plugin
+
+### Patch Changes
+
+- Updated dependencies [[`30c9913`](https://github.com/rabbitholegg/questdk-plugins/commit/30c99133337b9986d28081baad40186285ae0559)]:
+  - @rabbitholegg/questdk-plugin-utils@1.0.0-alpha.5
+
+## 1.0.0-alpha.2
+
+### Minor Changes
+
+- [#199](https://github.com/rabbitholegg/questdk-plugins/pull/199) [`e934b2d`](https://github.com/rabbitholegg/questdk-plugins/commit/e934b2da5427e038268d803f482df9c52dd052fe) Thanks [@Quazia](https://github.com/Quazia)! - Move questdk dependencies to root
+
 ## 1.0.0-alpha.1
 
 ### Minor Changes